Feat: 소비내역, 그래프 컴포넌트 분리
